refactor(checker): split reference type parameter helpers
